Deleted subject card popover that used ion popover and using custom one
